+++ trunk/emu-players/HesMapper.h
@@ -35,8 +35,8 @@ public:
 	void SetPsg(HuC6280Psg *psg) { mPsg = psg; }
 
 	virtual void Reset();
-	virtual uint8_t ReadByte(uint16_t addr);
-	virtual void WriteByte(uint16_t addr, uint8_t data);
+	virtual uint8_t ReadByte(uint32_t addr);
+	virtual void WriteByte(uint32_t addr, uint8_t data);
 
 private:
 	HuC6280Psg *mPsg;
