+++ trunk/emu-players/KssMapper.h
@@ -37,8 +37,8 @@ public:
 	virtual ~KssMapper();
 
 	virtual void Reset();
-	virtual uint8_t ReadByte(uint16_t addr);
-	virtual void WriteByte(uint16_t addr, uint8_t data);
+	virtual uint8_t ReadByte(uint32_t addr);
+	virtual void WriteByte(uint32_t addr, uint8_t data);
 
 	void SetAy(YmChip *ay) { mAy = ay; }
 	void SetScc(KonamiScc *scc) { mScc = scc; }
