+++ trunk/emu-players/SgcMapper.h
@@ -37,8 +37,8 @@ public:
 	void SetSystemType(uint8_t systemType);
 
 	virtual void Reset();
-	virtual uint8_t ReadByte(uint16_t addr);
-	virtual void WriteByte(uint16_t addr, uint8_t data);
+	virtual uint8_t ReadByte(uint32_t addr);
+	virtual void WriteByte(uint32_t addr, uint8_t data);
 	virtual uint8_t ReadPort(uint16_t port);
 	virtual void WritePort(uint16_t port, uint8_t addr);
