+++ trunk/emu-players/SidMapper.h
@@ -125,8 +125,8 @@ public:
 	void SetSidPlayer(SidPlayer *sidPlayer) { mSidPlayer = sidPlayer; }
 
 	virtual void Reset();
-	virtual uint8_t ReadByte(uint16_t addr);
-	virtual void WriteByte(uint16_t addr, uint8_t data);
+	virtual uint8_t ReadByte(uint32_t addr);
+	virtual void WriteByte(uint32_t addr, uint8_t data);
 
 	enum
 	{
